requirements.txt
# Core ML/DL
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Fine-tuning & Quantization
trl>=0.7.0
peft>=0.6.0
bitsandbytes>=0.41.0
scikit-learn>=1.3.0

fastapi 
uvicorn 
python-dotenv openai 
numpy

# API & App
openai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
streamlit>=1.28.0
pandas>=2.0.0

# Additional utilities
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0

# Logging & Monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Data validation
pydantic>=2.0.0
jsonschema>=4.19.0

# Environment management
python-dotenv>=1.0.0